+++ javascript/node/selenium-webdriver/test/proxy_test.js
@@ -111,7 +111,7 @@ test.suite(function(env) {
   }
 
   // Proxy support not implemented.
-  test.ignore(env.browsers(Browser.IE, Browser.OPERA, Browser.SAFARI)).
+  test.ignore(env.browsers(Browser.IE, Browser.SAFARI)).
   describe('manual proxy settings', function() {
     it('can configure HTTP proxy host', async function() {
       await createDriver(proxy.manual({
@@ -154,7 +154,7 @@ test.suite(function(env) {
 
   // PhantomJS does not support PAC file proxy configuration.
   // Safari does not support proxies.
-  test.ignore(env.browsers(Browser.IE, Browser.OPERA, Browser.SAFARI)).
+  test.ignore(env.browsers(Browser.IE, Browser.SAFARI)).
   describe('pac proxy settings', function() {
     it('can configure proxy through PAC file', async function() {
       await createDriver(proxy.pac(proxyServer.url('/proxy.pac')));